Add colors for project
The site is based on 2 colors; blue and yellow. Text is either white,
blue: base, or blue: dark. Buttons have a yellow background with blue:
dark text. the footer is also blue: dark.

The logo blue is different than the other blue. We may need to adjust
them to fit with the logo. Might be worth while to investigate altering
the hues to be more similar