Add button and ghost styling
The base styles for buttons as they appear in the mocks. This doesn't
really handle links or anything else, but it's a start.

Add button and ghost styling

The base styles for buttons as they appear in the mocks. This doesn't
really handle links or anything else, but it's a start.

Adds center text on buttons

Add block and narrow buttons