Add footer markup and data
This begins the footer for the site. Reorganizing it into 2 pieces. The
first is the colophon which contains the majority of the data, logo, and
call to action.

The second piece moves the parts of the copyright into a partial to keep

Add a base set of styles for the footer, as well as update the naming on the wrapper