Add collapse support for multiple Content Sections
The existing implementation would loop over each section creating a
service-list-id for each content item using a loop and the current
iterations index. When using multiple sections, the subsequent
sections would redefine the service-list-id from it's loop index, and not
the with the index of the whole count of content.

The easy solution is to use the content title as a id slug instead of
the loop item index.